+++ scvi/model/_scvi.py
@@ -3,9 +3,9 @@ import logging
 from anndata import AnnData
 
 from scvi._compat import Literal
-from scvi.compose._training import UnsupervisedTrainingMixin
 from scvi.dataloaders import AnnDataLoader
 from scvi.lightning import TrainingPlan
+from scvi.model.base._training_mixin import UnsupervisedTrainingMixin
 from scvi.modules import VAE
 
 from .base import ArchesMixin, BaseModelClass, RNASeqMixin, VAEMixin
